+++ app/template-customization-wizard/page.tsx
@@ -1,0 +1,5 @@
+import TemplateCustomizationWizard from "../../components/TemplateCustomizationWizard";
+
+export default function TemplateCustomizationWizardPage() {
+  return <TemplateCustomizationWizard />;
+}